Move pcan into own package
