+++ can/interfaces/pcan/__init__.py
@@ -1,0 +1,1 @@
+from can.interfaces.pcan.pcan import PcanBus
